+++ pkg/server/server.go
@@ -12,6 +12,9 @@ import (
 	"time"
 
 	"github.com/oklog/ulid/v2"
+	"github.com/openfga/openfga/internal/cachecontroller"
+
+	"github.com/openfga/openfga/pkg/storage/storagewrappers"
 
 	"github.com/openfga/openfga/internal/graph"
 
@@ -44,7 +47,6 @@ import (
 	serverconfig "github.com/openfga/openfga/pkg/server/config"
 	serverErrors "github.com/openfga/openfga/pkg/server/errors"
 	"github.com/openfga/openfga/pkg/storage"
-	"github.com/openfga/openfga/pkg/storage/storagewrappers"
 	"github.com/openfga/openfga/pkg/telemetry"
 	"github.com/openfga/openfga/pkg/typesystem"
 )
@@ -55,10 +57,9 @@ const (
 	AuthorizationModelIDHeader = "Openfga-Authorization-Model-Id"
 	authorizationModelIDKey    = "authorization_model_id"
 
-	ExperimentalEnableConsistencyParams ExperimentalFeatureFlag = "enable-consistency-params"
-	ExperimentalCheckOptimizations      ExperimentalFeatureFlag = "enable-check-optimizations"
-	ExperimentalAccessControlParams     ExperimentalFeatureFlag = "enable-access-control"
-	allowedLabel                                                = "allowed"
+	ExperimentalCheckOptimizations  ExperimentalFeatureFlag = "enable-check-optimizations"
+	ExperimentalAccessControlParams ExperimentalFeatureFlag = "enable-access-control"
+	allowedLabel                                            = "allowed"
 )
 
 var tracer = otel.Tracer("openfga/pkg/server")
@@ -148,6 +149,10 @@ type Server struct {
 
 	cacheLimit uint32
 	cache      storage.InMemoryCache[any]
+
+	cacheControllerEnabled bool
+	cacheControllerTTL     time.Duration
+	cacheController        cachecontroller.CacheController
 
 	checkQueryCacheEnabled bool
 	checkQueryCacheTTL     time.Duration
@@ -384,6 +389,20 @@ func WithCacheLimit(limit uint32) OpenFGAServiceV1Option {
 	}
 }
 
+// WithCacheControllerEnabled enables cache invalidation of different cache entities.
+func WithCacheControllerEnabled(enabled bool) OpenFGAServiceV1Option {
+	return func(s *Server) {
+		s.cacheControllerEnabled = enabled
+	}
+}
+
+// WithCacheControllerTTL sets the frequency for the controller to execute.
+func WithCacheControllerTTL(ttl time.Duration) OpenFGAServiceV1Option {
+	return func(s *Server) {
+		s.cacheControllerTTL = ttl
+	}
+}
+
 // WithCheckQueryCacheTTL sets the TTL of cached checks and list objects partial results
 // Needs WithCheckQueryCacheEnabled set to true.
 func WithCheckQueryCacheTTL(ttl time.Duration) OpenFGAServiceV1Option {
@@ -586,6 +605,10 @@ func NewServerWithOpts(opts ...OpenFGAServiceV1Option) (*Server, error) {
 
 		cacheLimit: serverconfig.DefaultCacheLimit,
 
+		cacheController:        cachecontroller.NewNoopCacheController(),
+		cacheControllerEnabled: serverconfig.DefaultCacheControllerEnabled,
+		cacheControllerTTL:     serverconfig.DefaultCacheControllerTTL,
+
 		checkQueryCacheEnabled: serverconfig.DefaultCheckQueryCacheEnabled,
 		checkQueryCacheTTL:     serverconfig.DefaultCheckQueryCacheTTL,
 
@@ -660,10 +683,17 @@ func NewServerWithOpts(opts ...OpenFGAServiceV1Option) (*Server, error) {
 		}
 	}
 
+	s.datastore = storagewrappers.NewCachedOpenFGADatastore(storagewrappers.NewContextWrapper(s.datastore), s.maxAuthorizationModelCacheSize)
+
 	if s.cacheLimit > 0 && (s.checkQueryCacheEnabled || s.checkIteratorCacheEnabled) {
 		s.cache = storage.NewInMemoryLRUCache([]storage.InMemoryLRUCacheOpt[any]{
 			storage.WithMaxCacheSize[any](int64(s.cacheLimit)),
 		}...)
+	}
+
+	if s.cache != nil && s.cacheControllerEnabled {
+		// TODO: replace checkQueryCacheTTL with checkIteratorCacheTTL once its introduced
+		s.cacheController = cachecontroller.NewCacheController(s.datastore, s.cache, s.cacheControllerTTL, s.checkQueryCacheTTL)
 	}
 
 	var checkCacheOptions []graph.CachedCheckResolverOpt
@@ -678,6 +708,7 @@ func NewServerWithOpts(opts ...OpenFGAServiceV1Option) (*Server, error) {
 	s.checkResolver, s.checkResolverCloser = graph.NewOrderedCheckResolvers([]graph.CheckResolverOrderedBuilderOpt{
 		graph.WithLocalCheckerOpts([]graph.LocalCheckerOption{
 			graph.WithResolveNodeBreadthLimit(s.resolveNodeBreadthLimit),
+			graph.WithOptimizations(s.IsExperimentallyEnabled(ExperimentalCheckOptimizations)),
 		}...),
 		graph.WithCachedCheckResolverOpts(s.checkQueryCacheEnabled, checkCacheOptions...),
 		graph.WithDispatchThrottlingCheckResolverOpts(s.checkDispatchThrottlingEnabled, checkDispatchThrottlingOptions...),
@@ -691,7 +722,6 @@ func NewServerWithOpts(opts ...OpenFGAServiceV1Option) (*Server, error) {
 		s.listUsersDispatchThrottler = throttler.NewConstantRateThrottler(s.listUsersDispatchThrottlingFrequency, "list_users_dispatch_throttle")
 	}
 
-	s.datastore = storagewrappers.NewCachedOpenFGADatastore(storagewrappers.NewContextWrapper(s.datastore), s.maxAuthorizationModelCacheSize)
 	s.checkDatastore = s.datastore
 
 	if s.cache != nil && s.checkIteratorCacheEnabled {
@@ -1071,6 +1101,7 @@ func (s *Server) Check(ctx context.Context, req *openfgav1.CheckRequest) (*openf
 		commands.WithCheckCommandLogger(s.logger),
 		commands.WithCheckCommandMaxConcurrentReads(s.maxConcurrentReadsForCheck),
 		commands.WithCheckCommandResolveNodeLimit(s.resolveNodeLimit),
+		commands.WithCacheController(s.cacheController),
 	).Execute(ctx, req)
 	if err != nil {
 		telemetry.TraceError(span, err)