+++ pkg/server/server_test.go
@@ -22,6 +22,7 @@ import (
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
+	"google.golang.org/protobuf/types/known/timestamppb"
 
 	"github.com/openfga/openfga/cmd/migrate"
 	"github.com/openfga/openfga/cmd/util"
@@ -691,7 +692,8 @@ func TestCheckDispatchThrottledTimeout(t *testing.T) {
 
 		type group
 			relations
-				define member: [user, group#member]
+				define other: [user]
+				define member: [user, group#member] or other
 		`)
 
 	writeAuthModelResp, err := s.WriteAuthorizationModel(context.Background(), &openfgav1.WriteAuthorizationModelRequest{
@@ -2121,7 +2123,8 @@ func TestServer_ThrottleUntilDeadline(t *testing.T) {
 
 		type group
 		relations
-			define member: [user, group#member]
+			define other: [user]
+			define member: [user, group#member, group#other]
 
 		type document
 		relations
@@ -2283,16 +2286,6 @@ func TestCheckWithCachedIterator(t *testing.T) {
 			relations
 				define viewer: [user, user:*, company#viewer]`).GetTypeDefinitions()
 
-	userTuple := &openfgav1.Tuple{
-		Key: tuple.NewTupleKey("company:1", "viewer", "user:1"),
-	}
-
-	usersetTuples := []*openfgav1.Tuple{
-		{
-			Key: tuple.NewTupleKey("license:1", "viewer", "company:1#viewer"),
-		},
-	}
-
 	mockController := gomock.NewController(t)
 	defer mockController.Finish()
 
@@ -2308,13 +2301,28 @@ func TestCheckWithCachedIterator(t *testing.T) {
 
 	mockDatastore.EXPECT().
 		ReadUserTuple(gomock.Any(), storeID, gomock.Any(), gomock.Any()).
-		Times(2).
-		Return(userTuple, nil)
+		AnyTimes().
+		DoAndReturn(
+			func(_ context.Context, _ string, tk *openfgav1.TupleKey, _ storage.ReadUserTupleOptions) (*openfgav1.Tuple, error) {
+				if tk.GetObject() == "company:1" {
+					return &openfgav1.Tuple{
+						Key:       tk,
+						Timestamp: timestamppb.Now(),
+					}, nil
+				}
+
+				return nil, storage.ErrNotFound
+			})
 
 	mockDatastore.EXPECT().
 		ReadUsersetTuples(gomock.Any(), storeID, gomock.Any(), gomock.Any()).
 		Times(1).
-		Return(storage.NewStaticTupleIterator(usersetTuples), nil)
+		Return(storage.NewStaticTupleIterator([]*openfgav1.Tuple{
+			{
+				Key:       tuple.NewTupleKey("license:1", "viewer", "company:1#viewer"),
+				Timestamp: timestamppb.Now(),
+			},
+		}), nil)
 
 	s := MustNewServerWithOpts(
 		WithDatastore(mockDatastore),